v2.0 1、完成从订单到生成费用的过程

+++ src/main/java/com/bangnd/sales/repository/AgentBusinessTypeRepository.java
@@ -3,5 +3,5 @@ package com.bangnd.sales.repository;
 import com.bangnd.sales.entity.*;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AgentBusinessTypeRepository extends JpaRepository<AgentBusinessType, Long> {
+public interface AgentBusinessTypeRepository extends JpaRepository<AgentBusinessType, Integer> {
 }
