+++ src/main/java/com/bangnd/sales/repository/AgentChannelTypeRepository.java
@@ -3,5 +3,5 @@ package com.bangnd.sales.repository;
 import com.bangnd.sales.entity.*;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AgentChannelTypeRepository extends JpaRepository<AgentChannelType, Long> {
+public interface AgentChannelTypeRepository extends JpaRepository<AgentChannelType, Integer> {
 }
