+++ src/main/java/com/bangnd/sales/service/AgentBusinessTypeService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface AgentBusinessTypeService {
     public List<AgentBusinessType> getAll();
 
-    public AgentBusinessType getAgentBusinessTypeById(long ind);
+    public AgentBusinessType getAgentBusinessTypeById(Integer ind);
 }
