+++ src/main/java/com/bangnd/util/repository/PeriodTypeRepository.java
@@ -3,5 +3,5 @@ package com.bangnd.util.repository;
 import com.bangnd.util.entity.PeriodType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PeriodTypeRepository extends JpaRepository<PeriodType, Long> {
+public interface PeriodTypeRepository extends JpaRepository<PeriodType, Integer> {
 }
